requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic-settings==2.2.1
celery[redis]==5.4.0
redis==5.0.4
google-auth==2.29.0
google-api-python-client==2.131.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
httpx==0.27.0
python-dotenv==1.0.1
python-multipart==0.0.9
jinja2==3.1.4
